Improve number parsing support: `0xcafe_babe` `123_321` etc.
Now supports the same kind of thousand-separators that Rust also uses.

+++ src/ast.rs
@@ -89,7 +89,7 @@ pub enum ExprLit<T> {
     BFE(BFieldElement),
     XFE(XFieldElement),
     Digest(Digest),
-    GenericNum(u64, T),
+    GenericNum(u128, T),
 }
 
 impl<T> BFieldCodec for ExprLit<T> {
